fix phpstan and phpcs

+++ Algorithm/Sort/BitonicSort.php
@@ -45,8 +45,8 @@ final class BitonicSort implements SortInterface
             return $list;
         }
 
-        $first  = self::sort(\array_slice($list, 0, $n / 2), SortOrder::ASC);
-        $second = self::sort(\array_slice($list, $n / 2), SortOrder::DESC);
+        $first  = self::sort(\array_slice($list, 0, (int) ($n / 2)), SortOrder::ASC);
+        $second = self::sort(\array_slice($list, (int) ($n / 2)), SortOrder::DESC);
 
         return self::merge(\array_merge($first, $second), $order);
     }
@@ -78,8 +78,8 @@ final class BitonicSort implements SortInterface
             }
         }
 
-        $first  = self::merge(\array_slice($list, 0, $n / 2), $order);
-        $second = self::merge(\array_slice($list, $n / 2), $order);
+        $first  = self::merge(\array_slice($list, 0, (int) ($n / 2)), $order);
+        $second = self::merge(\array_slice($list, (int) ($n / 2)), $order);
 
         return \array_merge($first, $second);
     }
